add get_stage_process_env to current_platform
